added first version of channel support
initial channel implementation

fixed docker compose and tox

fixed import

fixed mutichannel calls

removing autofailed test

fixed tests

changed travis testing framework for render-app

removing unneccesary environment variable

added simple channel unit test

fixed test

really fixed test

fixed travis bug

travis fix again..
