+++ tox.ini
@@ -8,6 +8,7 @@
 envlist = py27, py36
 
 [testenv]
-commands = pytest test
+commands = pytest
 deps =  -rrequirements.txt
         -rtest_requirements.txt
+passenv = RENDER_HOST RENDER_PORT RENDER_EXAMPLE_DATA RENDER_CLIENT_SCRIPTS RENDER_PYTHON_TEST_POOL_SIZE RENDER_JAVA_HOME
